separated useUserLocalStorage to it's own file which made it easier to manage
